# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags:
        - auth
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized
  /auth/register:
    post:
      operationId: register
      tags:
        - auth
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /clients:
    get:
      operationId: listClients
      tags:
        - clients
      summary: List clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
    post:
      operationId: createClient
      tags:
        - clients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
  /clients/{id}:
    get:
      operationId: getClient
      tags:
        - clients
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
    patch:
      operationId: updateClient
      tags:
        - clients
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
    delete:
      operationId: deleteClient
      tags:
        - clients
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
...
  /invoices:
    get:
      operationId: listInvoices
      tags:
        - invoices
      summary: List invoices
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
    post:
      operationId: createInvoice
      tags:
        - invoices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags:
        - invoices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
    patch:
      operationId: updateInvoice
      tags:
        - invoices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
    delete:
      operationId: deleteInvoice
      tags:
        - invoices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
...
  /events:
    get:
      operationId: listEvents
      tags:
        - calendar
      summary: List calendar events
      parameters:
        - name: start
          in: query
          schema:
            type: string
        - name: end
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalendarEvent'
    post:
      operationId: createEvent
      tags:
        - calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarEventInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEvent'
  /events/{id}:
    patch:
      operationId: updateEvent
      tags:
        - calendar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarEventUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEvent'
    delete:
      operationId: deleteEvent
      tags:
        - calendar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
    AuthResponse:
      type: object
      required:
        - user
        - token
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
    User:
      type: object
      required:
        - id
        - name
        - email
        - role
        - status
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          description: super_admin | manager | employee | client
        status:
          type: string
          description: active | suspended
        avatarUrl:
          type:
            - string
            - 'null'
        projectsCount:
          type:
            - integer
            - 'null'
        createdAt:
          type: string
        isActive:
          type: boolean
        roleId:
          type: integer
...
    Client:
      type: object
      required:
        - id
        - name
        - email
        - status
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        companyName:
          type:
            - string
            - 'null'
        email:
          type: string
        phone:
          type:
            - string
            - 'null'
        address:
          type:
            - string
            - 'null'
        gstNumber:
          type:
            - string
            - 'null'
        website:
          type:
            - string
            - 'null'
        industry:
          type:
            - string
            - 'null'
        status:
          type: string
          description: active | inactive | prospect
        notes:
          type:
            - string
            - 'null'
        createdAt:
          type: string
    ClientInput:
      type: object
      required:
        - name
        - email
        - status
      properties:
        name:
          type: string
        companyName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        gstNumber:
          type: string
        website:
          type: string
        industry:
          type: string
        status:
          type: string
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        companyName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        gstNumber:
          type: string
        website:
          type: string
        industry:
          type: string
        status:
          type: string
        notes:
          type: string
...
    Invoice:
      type: object
      required:
        - id
        - invoiceNumber
        - clientId
        - total
        - status
        - dueDate
        - createdAt
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        clientId:
          type: integer
        clientName:
          type:
            - string
            - 'null'
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
        subtotal:
          type: number
        tax:
          type: number
        discount:
          type: number
        total:
          type: number
        status:
          type: string
          description: draft | sent | paid | overdue | cancelled
        dueDate:
          type: string
        notes:
          type:
            - string
            - 'null'
        createdAt:
          type: string
    InvoiceLineItem:
      type: object
      required:
        - description
        - quantity
        - unitPrice
        - amount
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        amount:
          type: number
    InvoiceInput:
      type: object
      required:
        - clientId
        - items
        - dueDate
        - status
      properties:
        clientId:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
        tax:
          type: number
        discount:
          type: number
        status:
          type: string
        dueDate:
          type: string
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        clientId:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
        tax:
          type: number
        discount:
          type: number
        status:
          type: string
        dueDate:
          type: string
        notes:
          type: string
...
    CalendarEvent:
      type: object
      required:
        - id
        - title
        - startDate
        - type
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type:
            - string
            - 'null'
        startDate:
          type: string
        endDate:
          type:
            - string
            - 'null'
        type:
          type: string
          description: meeting | followup | deadline | invoice_due
        relatedId:
          type:
            - integer
            - 'null'
        createdAt:
          type: string
    CalendarEventInput:
      type: object
      required:
        - title
        - startDate
        - type
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        type:
          type: string
        relatedId:
          type: integer
    CalendarEventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        type:
          type: string
...