# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  /** super_admin | manager | employee | client */
  role: string;
  /** active | suspended */
  status: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  projectsCount?: number | null;
  createdAt: string;
  isActive?: boolean;
  roleId?: number;
}